Create README.md
This is a thirdperson survival game where the player needs to go around the map and collect a set number of logs to finish building a log house. Over time the characters hunger and thirst meter will
go down and to keep those meters up they need to find apples and water fountains in the level. If their thirst and hunger runs out, their health will slowy deplete and eventually die.